pyproject.toml
[project]
name = "rpsbattle"
version = "0.1.0"
description = "2D creature simulation playground for learning Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pygame-ce>=2.5.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[project.scripts]
rpsbattle = "sim.app:run"
